👂🚊 Checkpoint ./commands/inv.js:29068751/10 ./commands/clearinv.js:29068751/138 ./commands/addmoney.js:29068751/558 ./commands/setmoney.js:29068751/249

+++ commands/clearinv.js
@@ -2,11 +2,12 @@ const db = require('quick.db');
 const Discord = require('discord.js');
 
 module.exports = {
-    name: "inv",
+    name: "clearinv",
     description: "View your inventory",
 
 
     async run (client, message, args) {
+      if(!message.member.hasPermission("MANAGE_SERVER")) return message.channel.send('go fuck your self you dont have prem hahaha');
         let user = message.mentions.members.first() || message.guild.members.cache.get(args[0]);
         db.delete(user.id);
